Update Google auth to use the same sign in activity and show a spinner while loading
