+++ EmbeddedSocialClient/sdk/src/main/AndroidManifest.xml
@@ -30,17 +30,11 @@
 			android:label="@string/es_screen_title_signin"
 			android:windowSoftInputMode="stateHidden"
 			android:theme="@style/EmbeddedSocialSdkAppTheme.Stub">
-		</activity>
-		<activity
-			android:name="com.microsoft.embeddedsocial.ui.activity.GoogleCallbackActivity"
-			android:label="@string/es_screen_title_signin"
-			android:theme="@style/EmbeddedSocialSdkAppTheme.Stub">
 			<intent-filter>
 				<action android:name="com.microsoft.embeddedsocial.HANDLE_AUTHORIZATION_RESPONSE"/>
 				<category android:name="android.intent.category.DEFAULT"/>
 			</intent-filter>
 		</activity>
-
 		<activity
 			android:name="com.microsoft.embeddedsocial.ui.activity.PopularActivity"
 			android:configChanges="keyboard|keyboardHidden|screenSize"
